+++ package.json
@@ -63,6 +63,7 @@
     "@sentry/browser": "^5.6.3",
     "apexcharts": "^3.10.1",
     "axios": "0.19.0",
+    "babel-eslint": "^10.0.3",
     "babel-plugin-styled-components": "^1.10.7",
     "babel7-jest": "^1.0.1",
     "big.js": "^5.2.2",
@@ -90,7 +91,6 @@
     "jest-axe": "^3.2.0",
     "jest-changed-files": "^24.9.0",
     "jest-cli": "24.9.0",
-    "jquery": "^3.5.1",
     "keymaster": "1.6.2",
     "loadjs": "3.6.1",
     "lodash": "^4.17.19",
@@ -110,20 +110,11 @@
     "object-hash": "1.3.1",
     "path-to-regexp": "3.0.0",
     "pattern-library": "github:openstax/pattern-library#v1.1.0",
+    "perry-white": "0.1.0",
     "pluralize": "8.0.0",
     "popper.js": "1.15.0",
     "prop-types": "^15.7.2",
-    "prosemirror-example-setup": "^1.1.2",
-    "prosemirror-model": "^1.12.0",
-    "prosemirror-plus": "nathanstitt/prosemirror-plus#dist",
-    "prosemirror-schema-basic": "^1.1.2",
-    "prosemirror-schema-list": "^1.1.4",
-    "prosemirror-state": "^1.3.3",
-    "prosemirror-tables": "^1.1.1",
-    "prosemirror-view": "^1.16.1",
     "qs": "6.8.0",
-    "quill": "^1.3.7",
-    "quill-image-resize-alt-module": "^4.0.0",
     "rc-input-number": "^4.5.1",
     "rc-picker": "^1.4.12",
     "react": "16.9.0",
@@ -138,12 +129,10 @@
     "react-dnd-html5-backend": "9.3.4",
     "react-dnd-test-utils": "^9.3.4",
     "react-dom": "^16.9.0",
-    "react-dropzone": "^11.2.2",
     "react-floater": "^0.6.5",
     "react-is": "^16.13.1",
     "react-maskedinput": "4.0.1",
     "react-overlays": "^2.1.0",
-    "react-quill": "1.3.5",
     "react-router-dom": "5.1.2",
     "react-select": "^3.0.4",
     "react-sticky-table": "^5.1.4",
@@ -154,7 +143,6 @@
     "survey-react": "^1.1.7",
     "svg-url-loader": "3.0.0",
     "time-input-polyfill": "^1.0.7",
-    "trix": "^1.3.0",
     "twix": "1.3.0",
     "underscore": "1.9.1",
     "validator": "^11.1.0",
